# contracts/openapi/platform.yaml
openapi: 3.0.3
info:
  title: BuildingOS platform foundation API
  version: "0.1.0"
  description: >
    Exact operational routes for the BOS-001 platform foundation milestone.
    Service metadata plus BOS-010 F5a revenue APIs (subscription-service) and F2 building
    application & lifecycle APIs (building-service).
servers:
  - url: http://localhost:8080
    description: Local gateway
paths:
...
  /api/v1/platform/building-applications:
    get:
      summary: "Review queue (drafts never listed)"
      description: >
        meta carries page, size, total.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer }
        - name: size
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (status=DRAFT, size > 100)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/building-applications/{applicationId}/start-review:
    post:
      summary: "SUBMITTED to UNDER_REVIEW"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/building-applications/{applicationId}/request-information:
    post:
      summary: "UNDER_REVIEW to MORE_INFORMATION_REQUIRED (message shown to applicant)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/building-applications/{applicationId}/reject:
    post:
      summary: "UNDER_REVIEW to REJECTED (reason required)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/building-applications/{applicationId}/approve:
    post:
      summary: "Approve: creation fee settled, admin provisioned, building created in ONBOARDING"
      description: >
        D-12, D-26, D-29. Data is {application, buildingId, buildingStatus}. No subscription is attached.
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApproveRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (adminPhone, reason)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION, CREATION_FEE_UNPAID or FEE_NOT_CONFIGURED"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503":
          description: "DEPENDENCY_UNAVAILABLE (subscription-service or auth-service); nothing changed"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, traceId]
      properties:
        success: { type: boolean }
        data: {}
        meta: { type: object }
        traceId: { type: string }
...
    MessageRequest:
      type: object
      required: [message]
      properties:
        message: { type: string, maxLength: 1000 }
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, maxLength: 1000 }
    ApproveRequest:
      type: object
      required: [adminPhone, reason]
      properties:
        adminPhone: { type: string, description: Initial BUILDING_ADMIN; review UI pre-fills the applicant contact phone }
        reason: { type: string, maxLength: 1000 }
...
    ApiError:
      type: object
      required: [success, code, message, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code:
          type: string
          description: Stable error code; each operation lists the codes it may return.
          example: ACCESS_DENIED
        message: { type: string }
        traceId: { type: string, format: uuid }